move cmdhelper module under otaclient package
