+++ src/otaclient/app/boot_control/_common.py
@@ -23,7 +23,7 @@ from pathlib import Path
 from typing import Callable, Literal, Optional, Union
 
 from otaclient.app.configs import config as cfg
-from otaclient.app.boot_control import _cmdhelper as cmdhelper
+from otaclient import cmdhelper
 from otaclient_api.v2 import types as api_types
 from otaclient_common.common import (
     read_str_from_file,
