+++ src/otaclient/app/boot_control/_jetson_cboot.py
@@ -26,7 +26,7 @@ from pathlib import Path
 from typing import Generator, Optional
 
 from otaclient.app import errors as ota_errors
-from otaclient.app.boot_control import _cmdhelper as cmdhelper
+from otaclient import cmdhelper
 from otaclient.app.boot_control._common import (
     OTAStatusFilesControl,
     SlotMountHelper,
