+++ src/otaclient/cmdhelper.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Pre-defined subprocess calls."""
+"""Pre-defined subprocess calls, used by otaclient."""
 
 
 from __future__ import annotations
